requirements.txt
argparse
pillow
numpy
opencv-python
imageio